+++ lib/command/install.dart
@@ -2,6 +2,8 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
+import 'package:alfred_jetbrains_cli/assets/assets.dart';
+import 'package:alfred_jetbrains_cli/logger.dart';
 import 'package:args/command_runner.dart';
 import 'package:io/io.dart';
 import 'package:path/path.dart';
@@ -17,19 +19,38 @@ class InstallCommand extends Command<int> {
 
   @override
   FutureOr<int> run() {
+    consoleLogger.i('Install Workflow');
     // locate pref folder
     final pref = _locatePrefDirectory();
     // access workflows directory
     final workflows = Directory(join(pref.absolute.path, 'workflows'));
     // create folder if it doesn't exists (can happen on fresh install)
     workflows.createSync();
-    // prepare symbolic link
-    final destWorkflow = Link(join(workflows.absolute.path, workflowName));
+    // create workflow folder
+    final workflow = Directory(join(workflows.absolute.path, workflowName));
+    workflow.createSync();
 
-    // "install" (link) workflow
-    destWorkflow.existsSync()
-        ? destWorkflow.updateSync(Directory.current.absolute.path)
-        : destWorkflow.createSync(Directory.current.absolute.path);
+    // Write files to workflow folder
+    final infoPlistFile = File(join(workflow.absolute.path, 'info.plist'))
+      ..writeAsStringSync(infoPlist, mode: FileMode.writeOnly);
+    File(
+      join(workflow.absolute.path, 'icon.png'),
+    ).writeAsBytesSync(iconPng, mode: FileMode.writeOnly);
+    consoleLogger.i('Workflow installed: ${infoPlistFile.absolute.path}');
+    final String? binPath = _binPath();
+    if (binPath != null) {
+      consoleLogger.i('Move $binPath to workflow destination');
+
+      final Directory workflowBin = Directory(
+        join(workflow.absolute.path, 'bin'),
+      )..createSync();
+      File sourceFile = File(binPath);
+      // Move the bin to the destination
+      sourceFile.renameSync(
+        '${workflowBin.absolute.path}/${basename(sourceFile.absolute.path)}',
+      );
+    }
+    consoleLogger.i('Installation completed!');
     return ExitCode.success.code;
   }
 
@@ -39,5 +60,14 @@ class InstallCommand extends Command<int> {
     );
     final prefs = jsonDecode(prefsFile.readAsStringSync());
     return Directory(prefs['current']);
+  }
+
+  String? _binPath() {
+    if (Platform.executable.endsWith('/bin/dart') ||
+        Platform.executable.endsWith('\\bin\\dart.exe')) {
+      return null;
+    }
+
+    return Platform.executable;
   }
 }